requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.55.0
pydantic==2.10.1
pydantic_core==2.27.1
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
